Move backend Sentry observability into submodule
