+++ apps/backend/package.json
@@ -4,7 +4,8 @@
   "private": true,
   "scripts": {
     "dev": "npm run bundle --prefix ../../api && tsx watch src/index.ts",
-    "build": "npm run bundle --prefix ../../api && tsc",
+    "clean": "node -e \"require('node:fs').rmSync('dist', { recursive: true, force: true })\"",
+    "build": "npm run bundle --prefix ../../api && npm run clean && tsc",
     "start": "node dist/index.js",
     "lint": "tsc --noEmit",
     "test": "npm run bundle --prefix ../../api && node scripts/run-tests.mjs"
