+++ apps/backend/src/observability/sentry/index.test.ts
@@ -28,7 +28,7 @@ import {
   startBackendSpan,
   wrapBackendHandler,
   wrapBackendStreamHandler,
-} from "./sentry";
+} from ".";
 
 type FacadeTypeSample = Readonly<{
   service: BackendService;
